Integer literals go up to u128::MAX, not u64::MAX (#800)
`let x: i128 = 170141183460469231731687303715884105727` didn't compile. The
lexer parsed every integer literal into 64 bits, so anything above `u64::MAX`
came back as "invalid number literal" — which meant the 128-bit types worked for
arithmetic and not for the values that are the reason to reach for them. The
i128 test suite had to check results as *relationships* (`sum / 2 == a`) because
the expected number couldn't be written on either side of a comparison.

Tokens and AST nodes now carry an `i128`. That's the whole fix; the rest is
following it through the pipeline.

The one place a bit pattern is still needed is a `u128` above `i128::MAX` — the
single value range an `i128` can't represent. Everything below that is now its
own value, which took the `as u64 as i128` re-reads out of MIR lowering.

An unsuffixed literal gets a marker saying which band its magnitude landed in,
and the marker rules types *out* rather than picking one:

    let signed:   i128 = 100000000000000000000    // both compile
    let unsigned: u128 = 100000000000000000000

That's a behaviour change for the existing `u64`-by-magnitude marker too, which
used to pin the literal to `u64` outright. It meant this compiled and was false:

    let a: i128 = 9223372036854775807
    assert a + a == 18446744073709551614     // literal was a u64 → -2 in an i64

Both sides printed the same digits, so the assertion message read
"18446744073709551614 == 18446744073709551614". The literal takes the slot's
type now, and with nothing to constrain it the default is still the narrowest
type that holds it: i32, i64, u64, i128.

Diagnostics, which is where most of the work went:

- Past its slot, a literal names the range it missed rather than reading as a
  type mismatch: "integer literal `170141183460469231731687303715884105728` is
  out of range for `i128`", not "`u128` doesn't fit in `i128`".
- `u128`'s bounds are printed correctly. The range helper returned two `i128`s,
  so `u128`'s max came out as `i128::MAX` — off by half.
- A negative literal keeps its sign in the message instead of printing the
  magnitude it would have wrapped to.
- Digits past `u128::MAX` get their own message. "invalid number literal /
  unexpected character" sends people hunting for a typo that isn't there; it now
  says no integer type holds it, and names `u128::MAX`.
- Lexer errors carry their own code, label and "why" line. All of them were
  reported as E0001 "unexpected character", including the number ones.

Shift amounts: the checker types `a << b` with both sides the same, so a 128-bit
receiver gets a 128-bit shift count. The interpreter's 128-bit `shl`/`shr`/`pow`
demanded a plain `Value::Int` and rejected it. A count isn't a value, so it
widens back down, and out of `i64` range is out of any bit width anyway.

Comptime gained `I128`/`U128` values. Its arithmetic already ran in an `i128`,
so signed 128-bit is exact; a `u128` above `i128::MAX` has nowhere to be
computed and is refused rather than truncated (filed separately).

Semantic hashing feeds all 128 bits — two literals differing only above bit 64
hashed the same.

specs/types/primitives.md: L6 widens to `i128`, and new L8 says where digits
stop (`u128::MAX` up, `i128::MIN` down).

Tests: `tests/suite/t_i128_arithmetic.rk` keeps its relationship checks and
gains literal ones — the boundaries, the results the old tests could only
compare, radix/suffix/separator forms, slot-vs-magnitude typing, and 128-bit
Map keys (#663, which needed a writable wide literal to be exercised at all).
Two new compile-error fixtures cover the three slot-range failures and the two
ends nothing can hold.

+++ compiler/crates/rask-ast/src/expr.rs
@@ -16,7 +16,7 @@ pub struct Expr {
 #[derive(Debug, Clone)]
 pub enum ExprKind {
     /// Integer literal
-    Int(i64, Option<IntSuffix>),
+    Int(i128, Option<IntSuffix>),
     /// Float literal
     Float(f64, Option<FloatSuffix>),
     /// String literal
