+++ compiler/crates/rask-desugar/src/defaults.rs
@@ -639,7 +639,7 @@ mod tests {
     use rask_ast::Span;
 
     fn sp() -> Span { Span::new(0, 0) }
-    fn int_expr(v: i64) -> Expr {
+    fn int_expr(v: i128) -> Expr {
         Expr { id: NodeId(0), kind: ExprKind::Int(v, None), span: sp() }
     }
     fn str_expr(s: &str) -> Expr {
